+++ src/prism_sim/agents/allocation.py
@@ -1,4 +1,3 @@
-
 import numpy as np
 
 from prism_sim.constants import EPSILON